Rename original file
