+++ scripts/waypoint-ecs-add-tags.mjs
